requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
tqdm>=4.65.0
umap-learn>=0.5.3
scipy>=1.10.0

# Dashboard dependencies
flask>=2.3.0
pyarrow>=12.0.0
pillow>=9.5.0
requests>=2.31.0

# HuggingFace
huggingface-hub>=0.16.0
datasets>=2.14.0
transformers>=4.30.0

# V-JEPA2 dependencies
einops>=0.6.0
timm>=0.9.0

# Logging and visualization
tensorboard>=2.13.0
wandb>=0.15.0
seaborn>=0.12.0

# Development
jupyter>=1.0.0
ipython>=8.12.0
black>=23.0.0
flake8>=6.0.0
pytest>=7.4.0
torch>=2.0.0
numpy>=1.21.0
einops>=0.6.0
scipy>=1.7.0
matplotlib>=3.5.0
tqdm>=4.62.0
